requirements.txt
requests
prometheus_client
python-consul
pyyaml
